+++ daemon/daemon_windows.go
@@ -17,8 +17,8 @@ import (
 	"github.com/docker/docker/daemon/config"
 	"github.com/docker/docker/daemon/container"
 	"github.com/docker/docker/daemon/internal/libcontainerd/local"
+	"github.com/docker/docker/daemon/internal/libcontainerd/remote"
 	"github.com/docker/docker/daemon/network"
-	"github.com/docker/docker/libcontainerd/remote"
 	"github.com/docker/docker/libnetwork"
 	nwconfig "github.com/docker/docker/libnetwork/config"
 	winlibnetwork "github.com/docker/docker/libnetwork/drivers/windows"
