+++ libcontainerd/libcontainerd_windows.go
@@ -5,7 +5,7 @@ import (
 
 	containerd "github.com/containerd/containerd/v2/client"
 	"github.com/docker/docker/daemon/internal/libcontainerd/local"
-	"github.com/docker/docker/libcontainerd/remote"
+	"github.com/docker/docker/daemon/internal/libcontainerd/remote"
 	libcontainerdtypes "github.com/docker/docker/libcontainerd/types"
 	"github.com/docker/docker/pkg/system"
 )
